[fix]: 병합 오류 수정

+++ backend/src/signup/signup.service.ts
@@ -193,31 +193,12 @@ export class SignupService {
     }
   }
 
-  async deleteUser(checkToken: string) {
-    const token = checkToken.split(' ').pop();
+  async deleteUser(userId: number) {
 
-    const tokenPayload = await this.jwtService.verifyAsync(token);
 
     try {
-      if (
-        tokenPayload.isOAuthUser === true &&
-        tokenPayload.email_verified === true
-      ) {
-        await this.userSignupLogic.destroy({
-          where: {
-            email: tokenPayload.email,
-            isOAuthUser: tokenPayload.isOAuthUser,
-          },
-        });
-        console.log('Oauth 유저 회원 탈퇴함');
-        return;
-      } else if (tokenPayload.email && tokenPayload.isOAuthUser === false) {
-        await this.userSignupLogic.destroy({
-          where: { email: tokenPayload.email, phone: tokenPayload.phone },
-        });
-        console.log('개인회원 유저 회원 탈퇴함');
-        return;
-      }
+      await this.userSignupLogic.destroy({ where: { id: userId } })
+
     } catch (error) {
       throw new HttpException('회원탈퇴 실패', HttpStatus.FORBIDDEN);
     }
